code/requirements.txt
gitpython==3.1.0
logzero==1.7.0
requests==2.25.1
